// src/generated/MeshMetadata.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface MeshMetadata { bboxMin: Array<number> | null, bboxMax: Array<number> | null, triangleCount: number | null, volumeMm3: number | null, surfaceAreaMm2: number | null, computedAt: number, parseError: string | null, }